implemented tests for snippets api and login api. refactored snippets api endpoints to pass tests. implemented appropriate associations between database models (forgot to mention this on previos commits).

+++ controllers/snippets.js
@@ -13,12 +13,19 @@ snippetsRouter.get('/', async (req, res, next) => {
 snippetsRouter.post('/', async (req, res, next) => {
   const { title, content } = req.body;
 
+  if (!req.user) {
+    return res.status(401).send({
+      error: 'Need to authenticate to do that...tokens probably invalid',
+    });
+  }
+
   if (!(title && content)) {
-    return response.status(400).send({ error: 'invalid submission info' });
+    return res.status(400).send({ error: 'invalid submission info' });
   }
 
   try {
     const createdSnippet = await Snippets.create({ ...req.body });
+    await req.user.addSnippet(createdSnippet);
     res.status(201).send(createdSnippet);
   } catch (error) {
     next(error);
@@ -29,8 +36,19 @@ snippetsRouter.patch('/:id', async (req, res, next) => {
   const { content } = req.body;
   const id = req.params.id;
 
+  if (!content) {
+    return res.status(400).send({ error: 'invalid submission info' });
+  }
+
+  const snippet = await Snippets.findByPk(id);
+
+  if (snippet.userId !== req.user?.id) {
+    return res
+      .status(401)
+      .send({ error: 'You are not authenticated to do that' });
+  }
+
   try {
-    const snippet = await Snippets.findByPk(id);
     snippet.content = content;
     await snippet.save();
     res.status(204).end();
@@ -41,6 +59,14 @@ snippetsRouter.patch('/:id', async (req, res, next) => {
 
 snippetsRouter.delete('/:id', async (req, res, next) => {
   const id = req.params.id;
+
+  const snippet = await Snippets.findByPk(id);
+
+  if (snippet.userId !== req.user?.id) {
+    return res
+      .status(401)
+      .send({ error: 'You are not authenticated to do that' });
+  }
 
   try {
     await Snippets.destroy({ where: { id } });
